requirements.txt
aiohttp
python-dotenv
discord
google-generativeai